(nw) fixed the build

+++ src/mame/drivers/slapshot.cpp
@@ -635,14 +635,14 @@ void slapshot_state::driver_init()
 	u8 *dest = srcdata;
 	for (int c = 0; c < gx0->elements(); c++)
 	{
-		const u8 *c0base = gx0->get_data(c);
-		const u8 *c1base = gx1->get_data(c);
+		const u16 *c0base = gx0->get_data(c);
+		const u16 *c1base = gx1->get_data(c);
 
 		// loop over height
 		for (int y = 0; y < gx0->height(); y++)
 		{
-			const u8 *c0 = c0base;
-			const u8 *c1 = c1base;
+			const u16 *c0 = c0base;
+			const u16 *c1 = c1base;
 
 			for (int x = 0; x < gx0->width(); x++)
 			{
